Change free text annotation font size to float

+++ Zotero/Controllers/Database/Requests/EditAnnotationFontSizeDbRequest.swift
@@ -13,7 +13,7 @@ import RealmSwift
 struct EditAnnotationFontSizeDbRequest: DbRequest {
     let key: String
     let libraryId: LibraryIdentifier
-    let size: UInt
+    let size: CGFloat
 
     var needsWrite: Bool { return true }
